+++ tests/TestUtils.hs
@@ -1,31 +1,18 @@
-{-# LANGUAGE CPP, ForeignFunctionInterface #-}
+{-# LANGUAGE CPP #-}
+-- | Utility functions specific to 'directory' tests
 module TestUtils
   ( copyPathRecursive
-  , createSymbolicLink
   , modifyPermissions
-  , tryCreateSymbolicLink
+  , symlinkOrCopy
+  , supportsSymlinks
   ) where
 import Prelude ()
 import System.Directory.Internal.Prelude
 import System.Directory
-import System.FilePath ((</>))
+import System.FilePath ((</>), normalise, takeDirectory)
 #ifdef mingw32_HOST_OS
-import System.FilePath (takeDirectory)
+import System.Directory.Internal (win32_getFinalPathNameByHandle)
 import qualified System.Win32 as Win32
-#else
-import System.Posix (createSymbolicLink)
-#endif
-
-#ifdef mingw32_HOST_OS
-# if defined i386_HOST_ARCH
-#  define WINAPI stdcall
-# elif defined x86_64_HOST_ARCH
-#  define WINAPI ccall
-# else
-#  error unknown architecture
-# endif
-foreign import WINAPI unsafe "windows.h CreateSymbolicLinkW"
-  c_CreateSymbolicLink :: Ptr CWchar -> Ptr CWchar -> CULong -> IO CUChar
 #endif
 
 -- | @'copyPathRecursive' path@ copies an existing file or directory at
@@ -49,37 +36,61 @@ modifyPermissions path modify = do
   permissions <- getPermissions path
   setPermissions path (modify permissions)
 
-#ifdef mingw32_HOST_OS
-createSymbolicLink :: String -> String -> IO ()
-createSymbolicLink target link =
-  (`ioeSetLocation` "createSymbolicLink") `modifyIOError` do
-    isDir <- (fromIntegral . fromEnum) `fmap`
-             doesDirectoryExist (takeDirectory link </> target)
-    let target' = fixSlash <$> target
-    withCWString target' $ \ pTarget ->
-      withCWString link $ \ pLink -> do
-        status <- c_CreateSymbolicLink pLink pTarget isDir
-        if status == 0
-          then do
-            errCode <- Win32.getLastError
-            if errCode == c_ERROR_PRIVILEGE_NOT_HELD
-              then ioError . (`ioeSetErrorString` permissionErrorMsg) $
-                   mkIOError permissionErrorType "" Nothing (Just link)
-              else Win32.failWith "createSymbolicLink" errCode
-          else return ()
-  where c_ERROR_PRIVILEGE_NOT_HELD = 0x522
-        permissionErrorMsg = "no permission to create symbolic links"
-        fixSlash '/' = '\\'
-        fixSlash c   = c
-#endif
-
--- | Attempt to create a symbolic link.  On Windows, this falls back to
---   copying if forbidden due to Group Policies.
-tryCreateSymbolicLink :: FilePath -> FilePath -> IO ()
-tryCreateSymbolicLink target link = createSymbolicLink target link
+-- | On Windows, the handler is called if symbolic links are unsupported or
+-- the user lacks the necessary privileges to create them.  On other
+-- platforms, the handler is never run.
+handleSymlinkUnavail
+  :: IO a                               -- ^ handler
+  -> IO a                               -- ^ arbitrary action
+  -> IO a
+handleSymlinkUnavail _handler action = action
 #ifdef mingw32_HOST_OS
   `catchIOError` \ e ->
-    if isPermissionError e
-    then copyPathRecursive (takeDirectory link </> target) link
-    else ioError e
+    case ioeGetErrorType e of
+      UnsupportedOperation -> _handler
+      _ | isIllegalOperation e || isPermissionError e -> _handler
+      _ -> ioError e
+#endif
+
+-- | Create a symbolic link.  On Windows, this falls back to copying if
+-- forbidden by Group Policy or is not supported.  On other platforms, there
+-- is no fallback.  Also, automatically detect if the source is a file or a
+-- directory and create the appropriate type of link.
+symlinkOrCopy :: FilePath -> FilePath -> IO ()
+symlinkOrCopy target link = do
+  let fullTarget = takeDirectory link </> target
+  handleSymlinkUnavail (copyPathRecursive fullTarget link) $ do
+    isDir <- doesDirectoryExist fullTarget
+    (if isDir then createDirectoryLink else createFileLink)
+      (normalise target)
+      link
+
+supportsSymlinks :: IO Bool
+supportsSymlinks = do
+  canCreate <- supportsLinkCreation
+  canDeref <- supportsLinkDeref
+  return (canCreate && canDeref)
+
+-- | On Windows, test if symbolic link creation is supported and the user has
+-- the necessary privileges to create them.  On other platforms, this always
+-- returns 'True'.
+supportsLinkCreation :: IO Bool
+supportsLinkCreation = do
+  let path = "_symlink_test.tmp"
+  isSupported <- handleSymlinkUnavail (return False) $ do
+    True <$ createFileLink path path
+  when isSupported $ do
+    removeFile path
+  return isSupported
+
+supportsLinkDeref :: IO Bool
+supportsLinkDeref = do
+#ifdef mingw32_HOST_OS
+    True <$ win32_getFinalPathNameByHandle Win32.nullHANDLE 0
+  `catchIOError` \ e ->
+    case ioeGetErrorType e of
+      UnsupportedOperation -> return False
+      _ -> return True
+#else
+    return True
 #endif